created new test file for signUp mutation
