+++ api/tests/test_user_mutations.py
@@ -39,137 +39,6 @@ def user_schema_test_db_init():
 
         cleanup()
 
-
-def test_successful_creation(save):
-    """
-    Test that ensures a user can be created successfully using the api endpoint
-    """
-    result = run(
-        mutation="""
-        mutation{
-            createUser(displayName:"user-test", userName:"different-email@testemail.ca",
-                password:"testpassword123", confirmPassword:"testpassword123"){
-                user{
-                    userName
-                    displayName
-                }
-            }
-        }
-        """,
-    )
-
-    if "errors" in result:
-        fail("Tried to create a user, instead: {}".format(json(result)))
-
-    expected_result = {
-        "data": {
-            "createUser": {
-                "user": {
-                    "userName": "different-email@testemail.ca",
-                    "displayName": "user-test",
-                }
-            }
-        }
-    }
-
-    assert result == expected_result
-
-
-def test_email_address_in_use():
-    """Test that ensures each user has a unique email address"""
-    result = run(
-        mutation="""
-        mutation{
-            createUser(displayName:"testuser", userName:"testuser@testemail.ca",
-                password:"testpassword123", confirmPassword:"testpassword123"){
-                user{
-                    userName
-                }
-            }
-        }
-        """,
-    )
-
-    if "errors" in result:
-        fail("Error should create user, instead: {}".format(json(result)))
-
-    error_result = run(
-        mutation="""
-        mutation{
-            createUser(displayName:"testuser", userName:"testuser@testemail.ca",
-                password:"testpassword123", confirmPassword:"testpassword123"){
-                user{
-                    userName
-                }
-            }
-        }
-        """,
-    )
-
-    if "errors" not in error_result:
-        fail(
-            "Trying to create user with same username should fail, instead".format(
-                json(result)
-            )
-        )
-
-    [error] = error_result["errors"]
-    assert error["message"] == error_email_in_use()
-
-
-def test_password_too_short():
-    """
-    Test that ensure that a user's password meets the valid length requirements
-    """
-    result = run(
-        mutation="""
-        mutation{
-            createUser(displayName:"testuser", userName:"test@test-email.ca", password:"test", confirmPassword:"test"){
-                user{
-                    userName
-                }
-            }
-        }
-        """,
-    )
-
-    if "errors" not in result:
-        fail(
-            "Password too short when creating user should error out, instead: {}".format(
-                json(result)
-            )
-        )
-
-    [error] = result["errors"]
-    assert error["message"] == error_password_does_not_meet_requirements()
-
-
-def test_passwords_do_not_match():
-    """Test to ensure that user password matches their password confirmation"""
-    result = run(
-        mutation="""
-        mutation{
-            createUser(displayName:"testuser", userName:"test@test-email.ca", password:"A-Val1d-Pa$$word",
-                confirmPassword:"also-A-Val1d-Pa$$word"){
-                user{
-                    userName
-                }
-            }
-        }
-        """,
-    )
-
-    if "errors" not in result:
-        fail(
-            "Passwords do not match when creating user should error, instead: {}".format(
-                json(result)
-            )
-        )
-
-    [error] = result["errors"]
-    assert error["message"] == error_passwords_do_not_match()
-
-
 def test_update_password_success(save):
     """
     Test to ensure that a user is returned when their password is updated
